feat: Reduce the Discuss vote indicator to a per-card badge
Drops the "Your votes" recap panel and the "Only you" chip. Both were
noise: the recap restated on its own rows what the cards already say, and
"Only you" was redundant next to a card reading "1 votes" — the reader can
tell that nobody else backed it.

What stays is the signal itself: every topic card carries the reader's own
vote count ("Your 2 votes"), so the votes spent in the Vote phase remain
visible while the team works down the list, with a soft indigo border on
the cards they backed.

Removes components/session/MyVotesRecap.tsx and the
discussVoteSummary helper with it — the rank, distinct-backer and
support-scale derivations existed only for the recap. The remaining count
is one filter over the vote list the card already renders, inline in
DiscussPhase, so nothing is added to the session state.

Tests: discussPhaseMyVotesBadge covers the badge alone (stacked votes,
groups, unbadged topics, a reader who never voted, a topic with no vote
list, and that rendering writes no session state); the e2e spec now
asserts the Vote -> Discuss seam through the badges on both sides. The
CHANGELOG bullet is rewritten to describe what actually ships; VERSION
stays 29.0, one bump for the pull request.
